fix: rename main router and remove unused traits
