+++ src/main/scala/com/tudux/taxi/http/routes/MainRouter.scala
@@ -9,9 +9,7 @@ import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.DurationInt
 
 
-
-
-class TaxiStatsMainRouter(taxiTripActor: ActorRef)(implicit system: ActorSystem)
+class MainRouter(taxiTripActor: ActorRef)(implicit system: ActorSystem)
   {
 
   implicit val dispatcher: ExecutionContext = system.dispatcher
@@ -23,6 +21,8 @@ class TaxiStatsMainRouter(taxiTripActor: ActorRef)(implicit system: ActorSystem)
   val passengerRoutes = PassengerRoutes(taxiTripActor)
   val extraInfoRoutes = ExtraInfoRoutes(taxiTripActor)
   val serviceRoutes = ServiceRoutes(taxiTripActor)
+  val actorInfoRoutes = ActorInfoRoutes(taxiTripActor)
+  val pingRoutes = Ping()
 
   val routes: Route = {
     taxiTripRoutes.routes ~
@@ -30,7 +30,10 @@ class TaxiStatsMainRouter(taxiTripActor: ActorRef)(implicit system: ActorSystem)
     timeRoutes.routes ~
     passengerRoutes.routes ~
     extraInfoRoutes.routes  ~
-    serviceRoutes.routes
+    serviceRoutes.routes ~
+    actorInfoRoutes.routes ~
+    pingRoutes.routes
+
 
   }
 }
