+++ src/main/scala/com/tudux/taxi/http/routes/PassengerRoutes.scala
@@ -18,14 +18,7 @@ import spray.json._
 import scala.concurrent.ExecutionContext
 
 case class PassengerRoutes(taxiTripActor: ActorRef)(implicit system: ActorSystem, dispatcher: ExecutionContext,timeout: Timeout ) extends SprayJsonSupport
-  with TaxiCostStatProtocol
-  with TaxiTimeInfoStatProtocol
   with TaxiPassengerInfoProtocol
-  with TaxiExtraInfoProtocol
-  with CalculateDistanceCostProtocol
-  with CalculateAverageTripTimeProtocol
-  with GetAverageTipAmountProtocol
-  with GetTotalLoadedResponseProtocol
 {
   val routes: Route = {
       pathPrefix("api" / "yellowtaxi" / "passenger") {
