+++ src/main/scala/com/tudux/taxi/http/routes/Ping.scala
@@ -5,7 +5,7 @@ import akka.http.scaladsl.model.HttpResponse
 import akka.http.scaladsl.model.StatusCodes.OK
 import akka.http.scaladsl.server.Directives.{complete, get, path}
 
-case class Ping(system: ActorSystem) {
+case class Ping(implicit system: ActorSystem) {
   val routes = path("ping") {
     get {
       complete(HttpResponse(OK, entity = "pong"))
