+++ src/main/scala/com/tudux/taxi/http/routes/ServiceRoutes.scala
@@ -23,17 +23,11 @@ import spray.json._
 import scala.concurrent.{ExecutionContext, Future}
 
 case class ServiceRoutes(taxiTripActor: ActorRef)(implicit system: ActorSystem, dispatcher: ExecutionContext,timeout: Timeout )extends SprayJsonSupport
-  with TaxiCostStatProtocol
-  with TaxiTimeInfoStatProtocol
-  with TaxiPassengerInfoProtocol
-  with TaxiExtraInfoProtocol
   with CalculateDistanceCostProtocol
   with CalculateAverageTripTimeProtocol
   with GetAverageTipAmountProtocol
   with GetTotalLoadedResponseProtocol
 {
-
-
   val routes: Route = {
       pathPrefix("api" / "yellowtaxi" / "service" / "calculate-distance-cost") {
         get {
@@ -63,25 +57,6 @@ case class ServiceRoutes(taxiTripActor: ActorRef)(implicit system: ActorSystem, 
           complete(
             (taxiTripActor ? GetAverageTipAmount)
               .mapTo[GetAverageTipAmountResponse]
-              .map(_.toJson.prettyPrint)
-              .map(toHttpEntity)
-          )
-        }
-      } ~
-      pathPrefix("api" / "yellowtaxi" /  "actor" / "loaded") {
-        get {
-          val statTotalCostLoadedFuture: Future[Int] = (taxiTripActor ? GetTotalCostLoaded).mapTo[Int]
-          val statTotalPassengerInfoLoadedFuture: Future[Int] = (taxiTripActor ? GetTotalPassengerInfoLoaded).mapTo[Int]
-          val statTotalExtraInfoLoadedFuture: Future[Int] = (taxiTripActor ? GetTotalExtraInfoLoaded).mapTo[Int]
-          val statTotalTimeInfoFuture: Future[Int] = (taxiTripActor ? GetTotalTimeInfoInfoLoaded).mapTo[Int]
-          val totalLoadResponse = for {
-            r1 <- statTotalCostLoadedFuture
-            r2 <- statTotalExtraInfoLoadedFuture
-            r3 <- statTotalTimeInfoFuture
-            r4 <- statTotalPassengerInfoLoadedFuture
-          } yield LoadedStatsResponse(r1,r2,r3,r4)
-          complete(
-            totalLoadResponse.mapTo[LoadedStatsResponse]
               .map(_.toJson.prettyPrint)
               .map(toHttpEntity)
           )
