+++ src/main/scala/com/tudux/taxi/http/routes/TimeRoutes.scala
@@ -19,18 +19,9 @@ import spray.json._
 
 import scala.concurrent.ExecutionContext
 
-
 case class TimeRoutes(taxiTripActor: ActorRef)(implicit system: ActorSystem, dispatcher: ExecutionContext,timeout: Timeout ) extends SprayJsonSupport
-  with TaxiCostStatProtocol
   with TaxiTimeInfoStatProtocol
-  with TaxiPassengerInfoProtocol
-  with TaxiExtraInfoProtocol
-  with CalculateDistanceCostProtocol
-  with CalculateAverageTripTimeProtocol
-  with GetAverageTipAmountProtocol
-  with GetTotalLoadedResponseProtocol
 {
-
 
   def validateRequest[R: Validator](request: R)(routeIfValid: Route): Route = {
     validateEntity(request) match {
